Changes to migration
