+++ docker-compose.yml
@@ -23,6 +23,9 @@ services:
     build:
       context: .
       dockerfile: backend/database/migration/migration.Dockerfile
+      platforms:
+        - linux/amd64
+        - linux/arm64      
     depends_on:
       db:
         condition: service_healthy
